requirements.txt
fastapi
uvicorn
python-dotenv
openai
tavily-python
streamlit
requests
python-multipart
PyPDF2
scikit-learn
selenium
fpdf2